LayoutTests/media/media-source/media-source-real-remove-expected.txt

Tests SourceBuffer.remove() with real MP4 data. Removes [0, 2) and verifies that removal extends to the next sync frame boundary at PTS=4.

RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer('video/mp4; codecs="avc1.4d401e"'))
RUN(sourceBuffer.appendBuffer(initData))
EVENT(updateend)
RUN(sourceBuffer.appendBuffer(mediaData))
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.length == '1') OK
EXPECTED (sourceBuffer.buffered.start(0) == '0') OK
EXPECTED (sourceBuffer.buffered.end(0) == '8') OK
RUN(sourceBuffer.remove(0, 2))
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.length == '1') OK
EXPECTED (sourceBuffer.buffered.start(0) == '4') OK
END OF TEST
